- move utils to new file to cut down chance of circular dependency   - refactor tron api code, use eth rpc node for everything except balance check   - added more common util methods   - added methods to elastic module   - reduced debug logging   - wrapping error so it doesnt fill up the screen
