+++ src/abi/tron.ts
@@ -1,21 +1,14 @@
 
-import axios from "axios";
-import { ethers } from "ethers";
-import { fromHex, toHex } from 'tron-format-address'
 import { Address } from "../types";
-import convertResults from "./convertResults";
-import { debugLog } from "../util/debugLog"
-import { runInPromisePool, sliceIntoChunks, } from "../util"
 import { handleDecimals } from "../general";
 import pLimit from 'p-limit';
-import { getEnvRPC, getEnvValue } from "../util/env";
+import { getEnvValue } from "../util/env";
+import { evmToTronAddress, postJson, shortenString, sleepRandom, tronToEvmAddress } from "../util/common";
+import * as evmAbi from "./index";
 
 const limitRPCCalls = pLimit(+getEnvValue('TRON_RPC_CONCURRENCY_LIMIT', '5')!);
 
-const ownerAddress = 'TR7NHqjeKQxGTCi8q8ZY4pL8otSzgjLj6t'
-const MULTICALL_ADDRESS = 'TGXuuKAb4bnrn137u39EKbYzKNXvdCes98'
-
-const getEndpoint = () => getEnvRPC('tron') 
+const getEndpoint = () => getEnvValue('TRON_WALLET_RPC')
 
 type CallParams = any;
 
@@ -27,80 +20,18 @@ type CallOptions = {
 }
 
 export async function call(options: CallOptions) {
-  let { target, abi, params, isMulticall, } = options
-  target = unhexifyTarget(target)
-  const contractInterface = new ethers.Interface([abi]);
-  const functionABI = ethers.FunctionFragment.from(abi);
-  let inputTypes: any = (abi.inputs ?? []).map((i: any) => i.type)
-  if (isMulticall)
-    inputTypes = [ethers.ParamType.from({
-      components: [
-        { name: "target", type: "address" },
-        { name: "callData", type: "bytes" },
-      ],
-      name: "data",
-      type: "tuple[]",
-    })]
-  const outputTypes = (abi.outputs ?? []).map((i: any) => i.type)
-  params.forEach((v: any, i: number) => {
-    if (inputTypes[i] === 'address' && !v.startsWith('0x')) params[i] = toHex(v)
-  })
-  const callData = contractInterface._encodeParams(inputTypes, params)
-  const body = {
-    owner_address: ownerAddress,
-    contract_address: target,
-    function_selector: (contractInterface.fragments[0] as ethers.FunctionFragment).format(),
-    parameter: callData.slice(2),
-    visible: true,
-  }
-
-  const { constant_result: [data] } = await limitRPCCalls(() => post(body))
-
-  let decodedResult: any = contractInterface.decodeFunctionResult(
-    functionABI,
-    '0x' + data
-  );
-  decodedResult = [...decodedResult]
-  decodedResult.forEach((v: any, i: number) => {
-    if (outputTypes[i] === 'address' && v.startsWith('0x')) {
-      v = fromHex(v)
-      decodedResult[i] = v
-    }
-    let outputName = abi.outputs[i].name
-    if (outputName && outputName.length) decodedResult[outputName] = v
-  })
-  return {
-    output: convertResults(decodedResult, functionABI),
-  };
+  return evmAbi.call({ ...options, chain: 'tron', })
 }
 
 export async function multiCall(
-  functionABI: any,
-  calls: {
-    contract: string;
-    params: any[];
-  }[]
+  _functionABI: any,
+  _calls: { contract: string; params: any[]; }[]
 ) {
-  if (functionABI.inputs?.length) {
-    const inputs = functionABI.inputs
-    calls.forEach((call: any) => {
-      call.params?.forEach((v: any, i: number) => {
-        if (inputs[i].type === 'address') call.params[i] = hexifyTarget(v)
-      })
-    })
-  }
-  const returnValues = await executeCalls(functionABI, calls);
-
-  return returnValues.map((output: any, index: number) => {
-    return {
-      input: {
-        params: calls[index].params,
-        target: calls[index].contract,
-      },
-      success: output !== null,
-      output,
-    };
-  });
+  return evmAbi.multiCall({
+    abi: _functionABI,
+    calls: _calls.map(c => ({ target: c.contract, params: c.params })),
+    chain: 'tron',
+  })
 }
 
 
@@ -131,127 +62,30 @@ export async function getBalances(params: {
   };
 }
 
-async function executeCalls(
-  functionABI: any,
-  calls: {
-    contract: string;
-    params: any[];
-  }[]
-) {
-  const contractInterface = new ethers.Interface([functionABI]);
-  let fd = contractInterface.fragments[0] as ethers.FunctionFragment
-
-  const contractCalls = calls.map((call) => {
-    const data = contractInterface.encodeFunctionData(fd, call.params);
-    return {
-      to: call.contract,
-      data: data,
-    };
-  });
-
-  try {
-    const { output: [_, returnData] } = await call({ target: MULTICALL_ADDRESS, params: [contractCalls.map((call) => ({ target: hexifyTarget(call.to), callData: call.data, isMulticall: true }))], abi: MULTICALL_ABI, isMulticall: true })
-    return returnData.map((v: any) => {
-
-      let decodedResult: any = contractInterface.decodeFunctionResult(functionABI, v);
-      const outputTypes = (functionABI.outputs ?? []).map((i: any) => i.type)
-      let output = [...decodedResult]
-      output.forEach((v: any, i: number) => {
-        if (outputTypes[i] === 'address' && v.startsWith('0x')) {
-          v = fromHex(v)
-          output[i] = v
-        }
-        let outputName = functionABI.outputs[i].name
-        if (outputName && outputName.length) output[outputName] = v
-      })
-      return convertResults(output as ethers.Result, fd)
-    })
-  } catch (e) {
-    console.error(e)
-    if (calls.length > 10) {
-      const chunkSize = Math.ceil(calls.length / 5)
-      const chunks = sliceIntoChunks(calls, chunkSize)
-      debugLog(`Multicall failed, call size: ${contractCalls.length}, splitting into smaller chunks and trying again, new call size: ${chunks[0].length}`)
-      const response = await runInPromisePool({
-        items: chunks,
-        concurrency: 2,
-        processor: (calls: any) => executeCalls(functionABI, calls)
-      })
-      return response.flat()
-    }
-    debugLog("Multicall failed, defaulting to single transactions...")
-  }
-
-  return runInPromisePool({
-    items: calls,
-    concurrency: 3,
-    processor: async ({ contract, params }: any) => {
-      let result = null
-      try {
-        result = (await call({ target: contract, params, abi: functionABI })).output
-      } catch (e) {
-        debugLog(e)
-      }
-      return result
-    }
-  })
-}
-
-export function unhexifyTarget(address: string) {
-  if (address.startsWith('0x')) return fromHex(address)
-  return address
-}
-export function hexifyTarget(address: string) {
-  if (!address.startsWith('0x')) return toHex(address)
-  return address
-}
+export const unhexifyTarget = evmToTronAddress
+export const hexifyTarget = tronToEvmAddress
 
 async function post(body = {}, endpoint = '/wallet/triggerconstantcontract') {
-  const host = getEndpoint()
-  const headers: any = { 'Content-Type': 'application/json' }
-  const apiKey = getEnvValue('TRON_PRO_API_KEY')
-  if (apiKey) headers['TRON-PRO-API-KEY'] = apiKey
-  const { data } = await axios.post(host + endpoint, body, {
-    headers,
-  });
-  return data
-}
+  await sleepRandom(3000)   // sleep random time to avoid rate limit
+  const hosts = (getEndpoint()).split(',')
 
-// this is needed else correct function identifier wont be picked up while calling
-const MULTICALL_ABI = {
-  "inputs": [
-    {
-      "components": [
-        {
-          "internalType": "address",
-          "name": "target",
-          "type": "address"
-        },
-        {
-          "internalType": "bytes",
-          "name": "callData",
-          "type": "bytes"
-        }
-      ],
-      "internalType": "struct TronMulticall.Call[]",
-      "name": "calls",
-      "type": "tuple[]"
-    }
-  ],
-  "name": "aggregate",
-  "outputs": [
-    {
-      "internalType": "uint256",
-      "name": "blockNumber",
-      "type": "uint256"
-    },
-    {
-      "internalType": "bytes[]",
-      "name": "returnData",
-      "type": "bytes[]"
-    }
-  ],
-  "stateMutability": "view",
-  "type": "function"
-}
+  hosts.sort(() => Math.random() - 0.5) // shuffle hosts
 
+  const errors = []
+
+  for (const host of hosts) {
+    try {
+      const headers: any = { 'Content-Type': 'application/json' }
+      const apiKey = getEnvValue('TRON_PRO_API_KEY')
+      if (apiKey) headers['TRON-PRO-API-KEY'] = apiKey
+
+      const res = await postJson(host + endpoint, body, { headers, })
+      return res
+    } catch (e: any) {
+      // debugLog('Tron RPC error', e.message)
+      errors.push(e)
+    }
+  }
+
+  throw new Error(shortenString(`All TRON RPCs are not working. Errors: ${errors.map((i) => i.message).join('; ')}`, 1500))
+}
