+++ src/index.test.ts
@@ -119,7 +119,9 @@ test("imports", async () => {
         "addDebugLog": [Function],
         "addErrorLog": [Function],
         "addRuntimeLog": [Function],
+        "getAllLogs": [Function],
         "getClient": [Function],
+        "search": [Function],
         "writeLog": [Function],
       },
       "erc20": {
@@ -224,6 +226,8 @@ test("imports", async () => {
           "lookupBlock": [Function],
         },
         "convertToBigInt": [Function],
+        "evmToTronAddress": [Function],
+        "fetchJson": [Function],
         "formError": [Function],
         "formErrorString": [Function],
         "getProviderUrl": [Function],
@@ -234,13 +238,18 @@ test("imports", async () => {
         },
         "mergeBalances": [Function],
         "normalizeAddress": [Function],
+        "postJson": [Function],
         "removeTokenBalance": [Function],
         "runInPromisePool": [Function],
+        "shortenString": [Function],
+        "sleep": [Function],
+        "sleepRandom": [Function],
         "sliceIntoChunks": [Function],
         "sumChainTvls": [Function],
         "sumMultiBalanceOf": [Function],
         "sumSingleBalance": [Function],
         "tableToString": [Function],
+        "tronToEvmAddress": [Function],
       },
     }
   `);
